Split schemaIndex format into its own file
