+++ packages/dds/tree/src/feature-libraries/index.ts
@@ -44,7 +44,9 @@ export {
 export { ForestIndex } from "./forestIndex";
 export { singleMapTreeCursor, mapTreeFromCursor } from "./mapTreeCursor";
 export { buildForest } from "./object-forest";
-export { SchemaIndex, SchemaEditor, getSchemaString } from "./schemaIndex";
+export { SchemaIndex, SchemaEditor } from "./schemaIndex";
+// This is exported because its useful for doing comparisons of schema in tests.
+export { getSchemaString } from "./schemaIndexFormat";
 export {
 	singleStackTreeCursor,
 	CursorAdapter,
